requirements.txt
python-dotenv==1.0.1
mpmath==1.3.0
pathway[all]
pandas
requests
datetime
streamlit